form validation and html remodelling
I changed the old way the date was displayed to the one we were taught in the lab because the old way just had placeholders for the date and I needed actual values for the validator functions. I also created new javascript files(eventHandler.js, mainPrelogEventRegistration.js, signupEventRegistartion.js) I added code to  eventHandler.js and signupEvenRegistration.js for client-side validation of the fields in the form.