Remaining form validation
I made new files to handle events for the add content page and the blog post management page i also added code for the respective handlers and validators in eventhandler.js. I restructured the HTML files of those pages to make the comment box and post box into forms because their values will need to be submitted to a database in the future. I updated the signup page events and I changed the regexs of the password and username field, i also changed the signup validator function to show error text when the user tries to submit without entering any values in the fields. I changed some of the error text on the signup page to be more descriptive.